JCRVLT-180 implement filesystem based persistence layer * factoring in review comments ** abstracting out common methods of Registry Implementations ** persisting state in individual packagespecific and human readable files ** avoid holding state in memory ** cleaning up codebase & comments * adding support for external files

+++ vault-core/src/test/java/org/apache/jackrabbit/vault/packaging/integration/TestFSPackageRegistry.java
@@ -60,7 +60,11 @@ public class TestFSPackageRegistry extends IntegrationTestBase {
     @Before
     public void setUp() throws Exception {
         super.setUp();
-        FileUtils.cleanDirectory(DIR_REGISTRY_HOME);
+        if (DIR_REGISTRY_HOME.exists()) {
+            FileUtils.cleanDirectory(DIR_REGISTRY_HOME);
+        } else {
+            DIR_REGISTRY_HOME.mkdir();
+        }
         registry = new FSPackageRegistry(DIR_REGISTRY_HOME);
     }
